Rassembler les fonctions téléchargerNouveauxFichier* dans un même fichier
