+++ outils/sync-démarches-simplifiées-88444.js
@@ -19,8 +19,7 @@ import {isValidDate} from '../scripts/commun/typeFormat.js'
 //import checkMemory from '../scripts/server/checkMemory.js'
 
 import _schema88444 from '../data/démarches-simplifiées/schema-DS-88444.json' with {type: 'json'}
-import téléchargerNouveauxFichiersEspècesImpactées from './synchronisation-ds-88444/téléchargésNouveauxFichiersEspècesImpactées.js'
-import téléchargerNouveauxFichiersAP_AM from './synchronisation-ds-88444/téléchargerNouveauxFichiersAP_AM.js'
+import {téléchargerNouveauxFichiersEspècesImpactées, téléchargerNouveauxFichiersAP_AM, téléchargerNouveauxFichiersAttestations} from './synchronisation-ds-88444/téléchargerNouveauxFichiersParType.js'
 import { miseÀJourDécisionsAdministrativesDepuisDS88444 } from '../scripts/server/database/décision_administrative.js'
 
 /** @import {default as DatabaseDossier} from '../scripts/types/database/public/Dossier.ts' */
@@ -29,7 +28,7 @@ import { miseÀJourDécisionsAdministrativesDepuisDS88444 } from '../scripts/ser
 /** @import {default as RésultatSynchronisationDS88444} from '../scripts/types/database/public/RésultatSynchronisationDS88444.ts' */
 /** @import {default as Fichier} from '../scripts/types/database/public/Fichier.ts' */
 
-/** @import {DémarchesSimpliféesCommune, ChampDSCommunes, ChampDSDépartements, ChampDSRégions, Traitement, Message, ChampDSDépartement, DémarchesSimpliféesDépartement, ChampScientifiqueIntervenants, BaseChampDS} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
+/** @import {DémarchesSimpliféesCommune, ChampDSCommunes, ChampDSDépartements, ChampDSRégions, Message, ChampDSDépartement, DémarchesSimpliféesDépartement, ChampScientifiqueIntervenants, BaseChampDS} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
 /** @import {DossierDS88444, Annotations88444, Champs88444} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
 /** @import {SchemaDémarcheSimplifiée, ChampDescriptor} from '../scripts/types/démarches-simplifiées/schema.ts' */
 /** @import {DossierPourSynchronisation, DécisionAdministrativeAnnotation88444} from '../scripts/types/démarches-simplifiées/DossierPourSynchronisation.ts' */
@@ -673,6 +672,13 @@ const fichiersAP_AMTéléchargésP = téléchargerNouveauxFichiersAP_AM(
 
 
 
+/** Télécharger les nouveaux fichiers 'attestations' */
+/** @type {Promise<Map<DossierDS88444['number'], Fichier['id']> | undefined>} */
+const fichiersAttestionsTéléchargésP = téléchargerNouveauxFichiersAttestations(
+    dossiersDS,
+    laTransactionDeSynchronisationDS
+)
+
 
 /**
  * Synchronisation des dossiers
@@ -756,25 +762,27 @@ if(dossiersDS.length >= 1){
 
 /** Synchronisation des évènements de changement de phase */
 
-/** @type {Map<NonNullable<DatabaseDossier['id_demarches_simplifiées']>, Traitement[]>} */
+/** @type {Map<NonNullable<DatabaseDossier['id_demarches_simplifiées']>, Pick<DossierDS88444, 'traitements' | 'attestation'>>} */
 const évènementsPhaseDossierById_DS = new Map(dossiersDS.map(
-    ({id: id_DS, traitements}) => [id_DS, traitements])
+    ({id, traitements, attestation}) => [id, {traitements, attestation}])
 )
 
 
 let traitementsSynchronisés;
 
-/** @type {Map<DatabaseDossier['id'], Traitement[]>} */
-const idToTraitements = new Map()
+/** @type {Map<DatabaseDossier['id'], Pick<DossierDS88444, 'traitements' | 'attestation'>>} */
+const idToTraitementsAttestations = new Map()
 for(const [id_DS, traitements] of évènementsPhaseDossierById_DS){
     const dossierId = dossierIdByDS_id.get(id_DS)
 
-    //@ts-ignore
-    idToTraitements.set(dossierId, traitements)
+    if(!dossierId)
+        throw new Error(`Dossier.id manquant pour dossier DS numéro ${id_DS}`)
+
+    idToTraitementsAttestations.set(dossierId, traitements)
 }
 
-if(idToTraitements.size >= 1){
-    traitementsSynchronisés = dumpDossierTraitements(idToTraitements)
+if(idToTraitementsAttestations.size >= 1){
+    traitementsSynchronisés = dumpDossierTraitements(idToTraitementsAttestations, laTransactionDeSynchronisationDS)
 }
