+++ scripts/server/database/dossier.js
@@ -95,16 +95,16 @@ function traitementPhaseToDossierPhase(DSTraitementState){
 }
 
 /**
- * @param {Map<Dossier['id'], API_DS_SCHEMA.Traitement[]>} idToTraitements
+ * @param {Map<Dossier['id'], Pick<API_DS_SCHEMA.DossierDS88444, 'traitements' | 'attestation'>>} idToTraitementsAttestations
  * @param {import('knex').Knex.Transaction | import('knex').Knex} [databaseConnection]
  * @returns {Promise<any>}
  */
-export async function dumpDossierTraitements(idToTraitements, databaseConnection = directDatabaseConnection) {
+export async function dumpDossierTraitements(idToTraitementsAttestations, databaseConnection = directDatabaseConnection) {
     /** @type {ÉvènementPhaseDossier[]} */
     const évènementsPhaseDossier = [];
     
-    for(const [dossierId, apiTraitements] of idToTraitements){
-        for(const {dateTraitement, state, emailAgentTraitant, motivation} of apiTraitements){
+    for(const [dossierId, {traitements, attestation}] of idToTraitementsAttestations){
+        for(const {dateTraitement, state, emailAgentTraitant, motivation} of traitements){
             évènementsPhaseDossier.push({
                 phase: traitementPhaseToDossierPhase(state),
                 horodatage: new Date(dateTraitement),
